Extract validation functions to separate file
